kvm_local_deploy for Marvin
Implemented these new features:

* kvm_local_deploy.py -m ../marvin/mct-zone3-kvm3-kvm4.cfg
Deploy the hypervisors specified in Marvin config, kvm3 and kvm4 in this
example.

* kvm_local_deploy.py -m ../marvin/mct-zone3-kvm3-kvm4.cfg --force
Deploy the hypervisors specified in Marvin config, kvm3 and kvm4 in this
example. If they already exist, they will be destroyed first and then
newly created.

* kvm_local_deploy.py -m ../marvin/mct-zone3-kvm3-kvm4.cfg --delete
Delete the hypervisors specified in the Marvin config.

When this script finds multiple hypervisors to build, they are being
created in parallel to minimize waiting time.

The --delete and --force flags also work when deploying roles:

* kvm_local_deploy.py -r kvm -d 1 --delete
This will delete kvm1.

* kvm_local_deploy.py -r kvm -d 1 --force
This will delete kvm1 if it exists, then create a new kvm1 box.