Automatically add the NSX controller to Cosmic
I admit it's a bit hacky, but this was the easiest way to 'save' the zone uuid from NSX and add it when Cosmic was successfully deployed. Since Marvin doesn't support NSX controller addition and CloudMonkey is not setup on the base hv's it was easiest to directly insert it into the db.